+++ block/button/button.html
@@ -1,0 +1,6 @@
+<button class="button button--control-switch">
+    <svg class="button__icon" viewBox="0 -12 24 48">
+        <path d="M9 17l6 0l0 -3a6 6 0 1 0 -6 0z m1 4l4 0"/>
+    </svg>
+    <span class="button__title">Balcony</span>
+</button>
