+++ block/control/control.html
@@ -1,0 +1,7 @@
+<ul class="control">
+    <li class="control__item"></li>
+    <li class="control__item"></li>
+    <li class="control__item"></li>
+    <li class="control__item"></li>
+    <li class="control__item"></li>
+</ul>
