+++ block/label/label.html
@@ -1,0 +1,1 @@
+<h4 class="label">1st floor</h4>
