+++ block/nav/nav.html
@@ -1,0 +1,11 @@
+<nav class="nav"></nav>
+    <ul class="nav__list nav__list--quick"></ul>
+    <ul class="nav__list nav__list--filter"></ul>
+        <div class="nav__list nav__list--main-wrapper">
+        <ul class="nav__list nav__list--main">
+            <li class="nav__item nav__item--title">
+            </li>
+        </ul>
+    </div>
+</nav>
+
