+++ block/page/page.html
@@ -1,0 +1,6 @@
+<div class="page">
+    <section class="page__screen-saver"></section>
+    <section class="page__trap"></section>
+    <section class="page__control"></section>
+    <section class="page__nav"></section>
+</div>
