+++ css/main.css
@@ -1,10 +1,7 @@
-@import "reset-style.css";
+@import "common.css";
 
-@import "common.blocks/common.css";
-@import "common.blocks/fonts.css";
-@import "common.blocks/page.css";
-@import "common.blocks/label.css";
-@import "common.blocks/nav.css";
-@import "common.blocks/control.css";
-@import "common.blocks/button.css";
-@import "common.blocks/debug.css";
+@import "../block/page/page.css";
+@import "../block/label/label.css";
+@import "../block/nav/nav.css";
+@import "../block/control/control.css";
+@import "../block/button/button.css";
