--- css/reset-style.css
@@ -1,123 +1,0 @@
-* {
-    box-sizing: border-box;
-    padding: 0;
-    margin: 0;
-    border: 0;
-}
-
-html, body {
-    --page-bg-color: #212121;
-    width: 100%;
-    height: 100%;
-    background-color: var(--page-bg-color);
-    overflow: hidden;
-}
-
-body {
-    line-height: 1; 
-    text-rendering: optimizeLegibility;
-    text-decoration-skip: objects;
-    -webkit-text-size-adjust: 100%;
-    -webkit-font-smoothing: antialiased;
-    -webkit-tap-highlight-color: transparent;
-}
-
-a {
-    color: inherit;
-    text-decoration: none
-}
-
-ul,
-li {
-    /* padding: 0;
-    margin: 0; */
-    list-style-type: none;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-    /* margin: 0; */
-}
-
-:focus {
-    outline: none;
-}
-
-audio,
-video {
-    display: block;
-}
-
-button,
-input,
-optgroup,
-select,
-textarea {
-    /* padding: 0;
-    margin: 0; */
-    border: none;
-    border-radius: 0;
-    box-shadow: none;
-    background-color: transparent;
-    font: inherit;
-    color: inherit;
-    letter-spacing: inherit;
-}
-
-button {
-    overflow: visible;
-}
-
-button {
-    text-align: left;
-    text-transform: none;
-}
-
-button,
-[type='button'],
-[type='reset'],
-[type='submit'] {
-    cursor: pointer;
-    -webkit-appearance: none;
-}
-
-button::-moz-focus-inner,
-[type='button']::-moz-focus-inner,
-[type='reset']::-moz-focus-inner,
-[type='submit']::-moz-focus-inner {
-    border: none;
-    /* padding: 0; */
-}
-
-button:-moz-focusring,
-[type='button']:-moz-focusring,
-[type='reset']:-moz-focusring,
-[type='submit']:-moz-focusring {
-    outline: none;
-}
-
-[type='number']::-webkit-inner-spin-button,
-[type='number']::-webkit-outer-spin-button {
-    height: auto;
-}
-
-[type='search']::-webkit-search-decoration {
-    -webkit-appearance: none;
-}
-
-[type='search'] {
-    outline: none;
-}
-
-::-webkit-file-upload-button {
-    -webkit-appearance: button;
-    font: inherit; 
-}
-
-address {
-    font-style: normal;
-}
